resizeimageMoveHelpers
