+++ helpers/index.js
@@ -1,5 +1,6 @@
 const  HttpError  = require("./HttpError");
 const ctrlWrapper = require("./ctrlWrapper");
 const handleMongooseError = require("./handleMongooseError");
+const resizeImage = require("./resizeImage")
 
-module.exports = { HttpError, ctrlWrapper, handleMongooseError }
+module.exports = { HttpError, ctrlWrapper, handleMongooseError,  resizeImage}
